requirements_record.txt
fastapi[standard-no-fastapi-cloud-cli]
pyjwt
passlib[bcrypt]
pydantic-settings